fix: add schema for run config
fix: Add build command
refactor: adjust location of config definition
