+++ packages/cspell-tools/package.json
@@ -10,8 +10,10 @@
     "cspell-tools-cli": "bin.js"
   },
   "scripts": {
-    "build": "tsc -p .",
+    "build": "pnpm run build-schema && pnpm run compile",
     "build-dev": "tsc -p tsconfig.dev.json",
+    "build-schema": "ts-json-schema-generator --no-top-ref --path src/config/index.ts --type RunConfig --validation-keywords deprecated  -o  ./cspell-tools.config.schema.json",
+    "compile": "tsc -p .",
     "watch": "tsc -p . -w",
     "clean-build": "pnpm run clean && pnpm run build",
     "clean": "rimraf dist temp coverage .tsbuildinfo",
@@ -67,7 +69,8 @@
     "jest": "^29.1.2",
     "lorem-ipsum": "^2.0.8",
     "rimraf": "^3.0.2",
-    "shelljs": "^0.8.5"
+    "shelljs": "^0.8.5",
+    "ts-json-schema-generator": "^1.1.2"
   },
   "main": "bin.js"
 }
