+++ packages/cspell-tools/src/compiler/compile.test.ts
@@ -2,7 +2,7 @@ import * as path from 'path';
 import { spyOnConsole } from '../test/console';
 import { createTestHelper } from '../test/TestHelper';
 import { compile } from './compile';
-import { CompileRequest, Target } from './config';
+import { CompileRequest, Target } from '../config';
 import { readTextFile } from './readTextFile';
 
 const testHelper = createTestHelper(__filename);
