+++ packages/cspell-tools/src/compiler/createCompileRequest.ts
@@ -1,6 +1,6 @@
 import * as path from 'path';
 import { CompileCommonAppOptions } from '../AppOptions';
-import { CompileRequest, DictionaryFormats, Target } from './config';
+import { CompileRequest, DictionaryFormats, Target } from '../config';
 
 export function createCompileRequest(sources: string[], options: CompileCommonAppOptions): CompileRequest {
     const { max_depth, maxDepth, experimental, split, keepRawCase, useLegacySplitter } = options;
