+++ packages/cspell-tools/src/compiler/index.ts
@@ -1,4 +1,4 @@
 export { compileWordList, compileTrie } from './wordListCompiler';
 export { type Logger, setLogger } from './logger';
 export { compile, compileTarget } from './compile';
-export type { CompileRequest, CompileTargetOptions, RunConfig } from './config';
+export type { CompileRequest, CompileTargetOptions, RunConfig } from '../config';
