+++ packages/cspell-tools/src/config/config.ts
@@ -38,6 +38,11 @@ export interface CompileTargetOptions {
 
 export interface Target {
     /**
+     * Optional Name of the target.
+     */
+    name?: string;
+
+    /**
      * The target filename
      */
     filename: FilePath;
@@ -109,18 +114,4 @@ export interface SourceConfig {
      * @default false
      */
     keepRawCase?: boolean | undefined;
-}
-
-export function isFilePath(source: DictionarySource): source is FilePath {
-    return typeof source === 'string';
-}
-
-export function isFileSource(source: DictionarySource): source is FileSource {
-    if (!source || isFilePath(source)) return false;
-    return (<FileSource>source).filename !== undefined;
-}
-
-export function isFileListSource(source: DictionarySource): source is FileListSource {
-    if (!source || isFilePath(source)) return false;
-    return (<FileListSource>source).listFile !== undefined;
 }
