+++ packages/cspell-tools/src/config/index.ts
@@ -1,0 +1,12 @@
+export {
+    CompileRequest,
+    CompileTargetOptions,
+    DictionaryFormats,
+    DictionarySource,
+    FileListSource,
+    FilePath,
+    FileSource,
+    RunConfig,
+    SourceConfig,
+    Target,
+} from './config';
